requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
httpx==0.27.2
jinja2==3.1.4
anthropic==0.39.0
python-dotenv==1.0.1
python-multipart==0.0.12
